Fix import statement in SMS backends package
The __init__.py file in the SMS backends package previously
imported all of the backend modules to construct a dictionary
of backend classes. However, the async_aws.py module relied on
a third-party dependency called aioboto3 that is only installed
if the user specifies the async option when they install aldera.
Therefore, if the user installed aldera without the async option,
and tried to import the SMS backends package, a
ModuleNotFoundError was raised. This has been fixed.